requirements.txt
numpy~=1.18.1
torch
matplotlib~=3.3.0
kaggle
split-folders
torchvision~=0.9.0
pillow~=7.0.0
torchsummary
